Extracted functions from model.py into database.py and helper.py
